Add print() methods
So we can see the actual DAWG as an automaton.
Note that iterating over the state set of a DAWG is currently
a little tricky, since it will contain "deleted" states. You
have to check the free list to make sure it's valid.
This will be a problem for very large word lists, since
checking the free list is linear.