Register start state. Add id field to State.
Also added print() method to State, which simplifies the print
method for StateSet a lot.

Add id field to State. This will save a lot of passing around
both &State and its ID for things that need both.

Add state 0 to registry as part of finishing up. That means that
now the registry contains *all* (and only) states of the final
automaton. (Well, their IDs anyway.)