Hash increments appear to be working
Managed to get around borrow checker problems by making direct
accesses to the state vector and arc vectors within it, rather
than grabbing a reference to a State and holding onto it while
operating on its arcs and downstream states and so on.

Still pretty crude looking. Should be tidied up.

Also, we hold onto the raw StateSet and registry HashMap for
longer than we need them. Maybe transition from a StatesBuilder
to a HashBuilder? Then we can discard the data held by the
StatesBuilder, and just move the compressed state vector to
the HashBuilder and finally to the DAWG.